+++ src/envprobe/settings/variable_tracking.py
@@ -20,6 +20,22 @@ except ImportError:
     from envprobe.compatibility import nullcontext  # noqa
 
 
+K_DEFAULT_SETTING = "default"
+K_TRACK_LIST = "explicit_track"
+K_IGNORE_LIST = "explicit_ignore"
+
+
+def get_tracking_file_name():
+    """Returns the expected default name of the tracking configuration file.
+
+    Warning
+    -------
+    This method only returns the **filename** for the configuration file,
+    not its location or full path.
+    """
+    return "variable_tracking.json"
+
+
 class VariableTracking:
     """Allows access to the user's configuration files about which variables
     are tracked.
@@ -30,13 +46,13 @@ class VariableTracking:
     the local settings trumping the global ones.
     """
 
-    config_schema_global = {"default": True,
-                            "explicit_track": set(),
-                            "explicit_ignore": set()
+    config_schema_global = {K_DEFAULT_SETTING: True,
+                            K_TRACK_LIST: set(),
+                            K_IGNORE_LIST: set()
                             }
 
-    config_schema_local = {"explicit_track": set(),
-                           "explicit_ignore": set()
+    config_schema_local = {K_TRACK_LIST: set(),
+                           K_IGNORE_LIST: set()
                            }
 
     def __init__(self, global_configuration=None, local_configuration=None):
@@ -51,10 +67,12 @@ class VariableTracking:
         global_configuration : context-capable dict, optional
         local_configuration : context-capable dict, optional
         """
-        self._global = global_configuration if global_configuration else \
-            nullcontext(dict(self.config_schema_global))
-        self._local = local_configuration if local_configuration else \
-            nullcontext(dict(self.config_schema_local))
+        self._global = global_configuration \
+            if global_configuration is not None \
+            else nullcontext(dict(self.config_schema_global))
+        self._local = local_configuration \
+            if local_configuration is not None \
+            else nullcontext(dict(self.config_schema_local))
 
     @property
     def global_tracking(self):
@@ -62,13 +80,16 @@ class VariableTracking:
         **global** configuration.
         """
         with self._global as conf:
-            return conf.get("default", True)
+            return conf.get(K_DEFAULT_SETTING, True)
 
     @global_tracking.setter
     def global_tracking(self, new_value):
         """Sets the global tracking behaviour."""
         with self._global as conf:
-            conf["default"] = bool(new_value)
+            if new_value is None:
+                del conf[K_DEFAULT_SETTING]
+            else:
+                conf[K_DEFAULT_SETTING] = bool(new_value)
 
     @property
     def local_tracking(self):
@@ -76,13 +97,16 @@ class VariableTracking:
         **local** configuration.
         """
         with self._local as conf:
-            return conf.get("default", None)
+            return conf.get(K_DEFAULT_SETTING, None)
 
     @local_tracking.setter
     def local_tracking(self, new_value):
         """Sets the local tracking behaviour."""
         with self._local as conf:
-            conf["default"] = bool(new_value)
+            if new_value is None:
+                del conf[K_DEFAULT_SETTING]
+            else:
+                conf[K_DEFAULT_SETTING] = bool(new_value)
 
     @property
     def default_tracking(self):
@@ -92,19 +116,19 @@ class VariableTracking:
 
     def _unmark(self, variable_name, configuration):
         try:
-            configuration["explicit_track"].remove(variable_name)
+            configuration[K_TRACK_LIST].remove(variable_name)
         except KeyError:
             pass
         try:
-            configuration["explicit_ignore"].remove(variable_name)
+            configuration[K_IGNORE_LIST].remove(variable_name)
         except KeyError:
             pass
 
     def _tracked(self, variable_name, configuration):
-        return variable_name in configuration["explicit_track"]
+        return variable_name in configuration[K_TRACK_LIST]
 
     def _ignored(self, variable_name, configuration):
-        return variable_name in configuration["explicit_ignore"]
+        return variable_name in configuration[K_IGNORE_LIST]
 
     def _config(self, variable_name, configuration):
         return self._tracked(variable_name, configuration) or \
@@ -114,13 +138,20 @@ class VariableTracking:
         # A variable cannot be tracked and ignored at the same time.
         self._unmark(variable_name, configuration)
 
-        key_to_add_to = "explicit_track" if track else "explicit_ignore"
+        key_to_add_to = K_TRACK_LIST if track else K_IGNORE_LIST
         configuration[key_to_add_to].add(variable_name)
 
     def track_local(self, variable_name):
         """Marks the variable to be tracked in the **local** scope."""
         with self._local as conf:
             self._mark(variable_name, conf, True)
+
+    def is_tracked_local(self, variable_name):
+        """Returns whether the variable is set to be tracked in the
+        **local** configuration file.
+        """
+        with self._local as conf:
+            return self._tracked(variable_name, conf)
 
     def ignore_local(self, variable_name):
         """Marks the variable to be ignored in the **local** scope."""
@@ -138,6 +169,13 @@ class VariableTracking:
         """Marks the variable to be tracked in the **global** scope."""
         with self._global as conf:
             self._mark(variable_name, conf, True)
+
+    def is_tracked_global(self, variable_name):
+        """Returns whether the variable is set to be tracked in the
+        **global** configuration file.
+        """
+        with self._global as conf:
+            return self._tracked(variable_name, conf)
 
     def ignore_global(self, variable_name):
         """Marks the variable to be ignored in the **global** scope."""